#569 Fix: Cleaned up page-view and moved backup-cache to a react hook
